Pivot hard and switch to elixir
With the recurring problems of ReasonML, a realization of a different
approach, and the possibility of Lumen filling in a gap in the future, I
am making the decision to switch to Elixir.

The MVP of the new approach is to build an index of tags to notes as a
markdown file and be able to all be stored on disk and commitable to a
git repo.

Wiring it directly into git could be a good next step. A more
full-featured editing/management experience is a subject to be
investigated later so that some note organization can be done more
immediately.